add cageArea add fan50AirSpeed

+++ src/main/java/com/toxa/ventilation/Data/Storage.java
@@ -11,8 +11,8 @@ public class Storage implements Serializable{
         this.description = description;
     }
 
-    public int getCapacity() {
-        return (int)capacity;
+    public double getCapacity() {
+        return capacity;
     }
 
     public void setCapacity(double capacity) {
